Correct labels for yearqtr index (fix #75)
Special case handling for creating labels when the index class is
yearqtr because zoo::format.yearqtr doesn't accept a "%b" format in
zoo 1.7.12.

This is not a bug in format.yearqtr. It behaves as documented and is
consistend with base R format methods. Adding support for higher
resolution formats would be awkward (should they be left-aligned,
centered, right-aligned, etc).