Throw error if j not in colnames (fix #48)
Make [.xts consistent with [.zoo and [.matrix by throwing an error if j
is character and not one of the object column names. [.xts would return
a malformed xts object if no values in j were found, and it would return
just the found values of j if only some columns were not found.