Allow indexTZ=GMT for index classes w/o TZ (#66)
Although it's not technically correct (since GMT *is* a timezone)
indexTZ="GMT" should work for all practical purposes. So check.TZ
should not warn in this case.

Also make the warning for indexTZ="" more explicit by printing empty
quotes instead of an empty string (i.e. from "...but indexTZ is " to
"...but indexTZ is ''").