Accept bty and text.col in addLegend (fix #95)
bty and text.col are not formal arguments to addLegend, so they reside
in ... and get passed to plot_legend. This causes them to appear in the
internal call to legend twice (once in our hard-coded call, and once in
...).

Add bty and text.col as formal arguments to plot_lines and check if
they're missing before assigning our default values.