Avoid rbind heap overflow (fix #118)
AddressSanitizer reports a heap overflow at line 216 when two objects
share an index value and the shared value happens to be the last value
in the x index.

This is because xp is one-based and when xp=nrow(x), real_xindex[xp-1]
returns the last element, and real_xindex[xp] accesses memory outside
the bounds of the x index.

The solution is to ensure xp < nrow(x) before checking to see if the
next x index value is different from the current x index value.