tzone=UTC if index<- value is class w/o TZ
The tzone attribute on the index was not being set to UTC when
index<-.xts was called and 'value' did not have a tzone attribute.

This could cause issues in to.monthly/quarterly/yearly if indexAt is a
class that does not have a timezone and drop.time=TRUE. This is because
xts:::.drop.time calls index(x) <- index(x). index<-.xts sets the
index's tzone attrbute to the tzone attribute of index(x), which
doesn't exist.

Check if 'value' is one of the classes without a timezone attribute, and
set the index tzone attribute to "UTC" if true.

Fixes #148.