tzone=UTC if indexClass<- value is class w/o TZ
A user could use indexClass(x) <- value to set the index class of 'x' to
a class that does not have a timezone, but the index tzone attribute
would not be set to UTC like it should.

Check if 'value' is one of the classes without a timezone attribute, and
set the index tzone attribute to "UTC" if true. Add a indexTZ = UTC
check to all indexClass tests.