Ensure to.frequency endpoints contain 0/nrow(x)
to.frequency() calls the toPeriod C function, which expects the
second argument to be the output of endptions(). endpoints() always
returns a vector where the first element is zero and the last element
is the length of the input.

If the first element is not zero, then the output of toPeriod will not
contain the aggregation for the first aggregation group.

Fixes #150.