Evaluate panels in parent frame of plot.xts call
A default call to  eval() evaluates the expression in the environment
from which eval() was called (i.e. the plot.xts environment in this
case). If an object in the expression does not exist in the plot.xts
environment, or in an environment attached to the search path, the
object will not be found.

This is because variable lookup will progress from the plot.xts
environment, to its parent environment (i.e. the environment plot.xts
was *created* in), the xts namespace. From there, lookup will continue
through the search path. If the object is not in an environment on the
search path, lookup will fail.

A simple example to illustrate the problem is calling plot.xts in a
function and using an object defined in the function in the panels
string.

foo <- function() {
  bar <- .xts(10 + cumsum(rnorm(100)), 1:100)
  plot(bar, panels='addSeries(rollmeanr(bar, 50), on=1, col=2)')
}
foo()  # Error in rollmean(..., align = align) : object 'bar' not found

Also remove if-statement, since both the true and false sections were
the same.