Special case when 'f' is length-1 character
Calling endpoints() when 'f' is a character vector longer than one
element will result in an error. There won't be an error if the first
element happens to match match one of the 'on' arguments endpoints()
accepts. But there will be a warning and the user will probably be
surprised by the output.

So, split.xts() should only call endpoints() if 'f' is a character
vector with 1 element. Otherwise, dispatch to split.zoo().

Fixes #134.