Correct endpoints for sparse daily data
endpoints() uses POSIXlt elements to determine the period breakpoints
for some values of 'on' (quarters, months, and days). This can be a
problem for sparse data, when there are contiguous observations that
have the same POSIXlt element used for aggregation, but different
years.

For each of the 3 cases where POSIXlt elements are used, ensure the
endpoints always vary by year, even if the POSIXlt element being used
does not vary across observations.

Fixes #169.