Check numeric index for non-finite index values
The check for finite index values was very early in the xts
constructor, which could cause problems if order.by wasn't strictly
a numeric vector (e.g. if it was POSIXlt--a list). This was to ensure
the index doesn't contain +/-Inf, NA, or NaN (see #173).

Move the check until after the index has been converted to
POSIXct/numeric, so we can be certain it is a numeric vector.

Fixes #194.