Fix protection calls and counters
There were several protect calls in do_merge_xts() that did not
increment the function's protect counter. These might not have been a
problem, but the rchk tool flagged the function as having an
unbalanced protect stack. Consistently using the protect counter makes
the function easier to understand.

The CAR(args) calls in mergeXts() do not need to be protected, because
they reference the args input parameter. Use PROTECT_WITH_INDEX() and
REPROTECT() inside loops to avoid the potential case of overflowing
the protect stack and causing R to error. As with do_merge_xts(),
consistently use the protect counter and avoid stand-alone UNPROTECT()
calls.

See #178.