Attempt to improve scalar subset performance
Subsetting xts objects is generally similar to (or faster than) that
for a similar matrix object, except in the case of extracting a single
element.

This is very much experimental, as the C function does no checking and
can bring down your entire R session. Still need to ensure all
attributes are copied to the new object (e.g. dims, names), but that
shouldn't be significant enough to erode the ~10x speedup.

  nr <- 1e6
  nc <- 50
  m <- matrix(1, nr, nc)
  x <- .xts(m, seq(nr))
  set.seed(21)
  i <- sample(nr, 1)
  j <- sample(nc, 1)
  fast <- function(x, i, j) {
    .Call("xts_ExtractScalar", x, i, j, PACKAGE="xts")
  }
  microbenchmark(m[i,j], x[i,j], fast(x,i,j), unit="us")

  Unit: microseconds
            expr    min      lq  median      uq     max neval
         m[i, j]  0.206  0.3580  0.5285  0.5665   1.166   100
         x[i, j] 32.764 33.5235 33.8025 34.3015 306.162   100
   fast(x, i, j)  1.758  2.0300  2.2960  2.5170  20.854   100