Set y-axis labels independent of periodicity
The y-axis labels were being set based on a portion of the distance
between actual data observations. If the distance between observations
was large, the labels could be off the plot region. If the distance
between observations was small, the labels could overlap the axis.

Use the `offset` and `pos` arguments to `text()` to position the axis
labels appropriately and independent of the periodicity of the data.

Fixes #85.