Create ylim when series is constant
plot.xts() would error if called on an object with a series of one
constant value (e.g. .xts(rep(1,10), 1:10)) because the y-coordinates
are the same.

Create different y-coordinates that are +/-20% from the series value.

Also move the check for all ylim equal to zero outside of the if/else
block. This avoids code duplication, and therefore potential future
bugs.

Fixes #156.