Use int64_t to avoid 32-bit long on Windows
The long C type is a 32-bit integer on Windows (even 64-bit Windows),
but a 64-bit integer on widely-used unix-alikes.  The C99 standard
includes a 64-bit integer, using the int64_t type defined in stdint.h.

Cast to int64_t instead of long when a 64-bit integer is needed.  Also
convert and rename real_tmp to int64_tmp to avoid a re-cast back to
double.  This should also make the logical comparisons more efficient.

See #202.