Avoid redundant calculations in endpoints
The calculations for the prior period's temporary value is always
redundant, because it had been calculated during the prior iteration.

Initialize the prior value before starting the loop.  Then copy the
current value to the prior value after comparing the two values and
setting the endpoint and before continuing to the next iteration.

This approach is ~2x faster than the original approach.