Ensure fill argument has at least one value
It's possible to set the fill argument to a zero-length vector, or
NULL (which is a zero-length vector when coerced to the common type of
all other inputs).  This is confusing and potentially alarming because
it looks as if random memory is being returned.

Check and ensure fill has at least one element before coercing it to
the common type.

Fixes #201.