Make as.matrix on zero-width xts similar to zoo
Calling as.matrix() on a zero-width xts object results in an error
because dimnames cannot be set on a non-array (i.e. an object without
dims).  as.matrix.zoo() has special cases for zoo objects that are
missing one or more dims.

Refactor as.matrix.xts() to follow the pattern in as.matrix.zoo(), so
the two functions behave similarly.

Fixes #130.