Evaluate par in parent environment
Calling eval() means we look for the par objects in plot.xts()'s
environment, which is the xts namespace (function environments are that
in which they were defined, not called).  This leads to "object not
found" errors.

Use eval.parent() to evaluate the par objects in the environment from
which plot.xts() was called.