Add new_environment() function
This function provides a way to create a new environment with 'Env' as
the parent environment.

The thought is that it will help encapsulate all actions associated
with a specific plot object instance. This isn't an issue now, because
there is only one global plot object instance. But it will be an issue
once we try to allow users to maintain multiple plot objects on
different devices.