Capture values from caller, instead of copying
Evaluate objects in the calling environment, so we don't need to copy
objects to 'lenv'. This prevents scoping issues if the caller passes
a named object instead of a value.

This isn't a problem when plotting functions are called like:

  addSeries(x, main = "My plot")

But may cause an 'object not found' error if a user calls a plot
function via:

  title <- "My plot"
  # 'title' may not be found when plot is drawn
  addSeries(x, main = title)