Capture and reset par values
We were changing values of par() without resetting them to their
original values before returning from plot.xts().

We changed 'bg' and then set it to our desired value again before
returning. We changed 'xpd' and attempted to reset it, but the 'oxpd'
object is a list and we need to explicitly reference the 'xpd' element
(oxpd$xpd).

We were capturing and resetting the previous value of 'cex', but it was
a little confusing because of the second call to par(cex=x$Env$cex*1.5)
in the if-statement for an X11 device. 'cex' is now set in one call.

Fixes #217.