Avoid floating-point precision diffs in tests
There is a potential for differences due to floating-point arithmetic
because date/time classes in R can be doubles, not just integers (even
Date can be double). So checking with identical() can lead to false
test failures.

Use the checkEqualsNumeric() comparison function to account for these
potential precision differences.

Fixes #219.