Ensure series has same index values as main panel
Users could trigger an error by trying to plot() and then add a series
using data with different index values. An error would occur if the
new data did not have the exact same number of rows as the main panel
data. Even if it happened to have the same number of rows, there is
nothing to guarantee the two series would be aligned by time.

Merge the value of 'x' in chart.lines() with the subset data from the
main panel of the current plot. Use a right-join to ensure that only
observations with the same timestamps as the main plot are returned.

This could still be a bit confusing, since the 'x' data may have
observations within the timespan of the main panel data, but not have
any observations on exactly the same timestamps. We could interpolate
the data for lines(), but that wouldn't make sense for points(). Maybe
it would be best to simply throw a warning.

Fixes #216.