Do not over-register as.Date.numeric method
Registering an S3 method that is the same as a S3 method from a
"standard" package is not allowed by CRAN. The function can remain
as-is, and we might even be able to export it, but leave it unexported
for now. The method will still be found for all internal xts functions
that depend on it.