Add explicit tests for isOrdered()
Some of this functionality was tested via the constructor tests. UBSAN
flagged some issues in do_is_ordered(), and these tests should trigger
similar issues if they exist at other locations in that function.

All tests should pass in the current version of the code. Note that
some behavior is not consistent (e.g. NA in the middle of integer
vectors does not produce the same results as NA in the middle of
double vectors).  Those should be addressed later.