Avoid undefined arithmetic w/NA_INTEGER (INT_MIN)
UBSAN flagged subtraction that included NA_INTEGER (the smallest signed
int). We're not testing for exact equality or using a tolerance, so we
should be able to do a simple logical comparison between the two
elements.

We have to be careful about NA_INTEGER because it's the smallest signed
int in C, but sorts as the max int in R. So we know the series is not
in increasing (decreasing) order if the first (last) element is
NA_INTEGER. Conversely, a series may still be in  in increasing
(decreasing) order if the last (first) element is NA_INTEGER.