Use location map to copy index to result
We have locations of the input rows and where they should be in the
result object. We use that map to copy coredata, but we can also use
it to copy the index. We need to do a few things before we can use
memcpy() to set the index values for the result.

We need to move the index type check, comparison function assignment,
and index missing values check to before the loop to determine map
locations. Use the compare_indexes() function inside the mapping loop,
so both index types can be used. Delete the old code for integer index
types.

Remove all references to, and required by, the set_index_* functions.
They're no longer needed now that the index is being set via memcpy().