Fix mapping logic for inner joins
When we are beyond the end of x (y), we should only set nodes and
increment locations for y (x) when right_join (left_join) is TRUE.
Otherwise break the loop, since we're out of observations for x (y)
and we're not including all observations from y (x).

Change all increment commands from `foo = foo+1` to the standard
foo++. The latter were there because the algorithm prototype was done
in R.

Also add unit tests for full-outer and full-inner joins for all data
types and both index types.