Make coredata() on zero-length xts same as zoo
coredata.zoo() simply removes the zoo attributes (e.g. index) and
returns the underlying matrix or vector. But coredata.xts() would
always return a zero-length vector the same type as the underlying
matrix.

Commit 779e6a0a50814d2456dd1b11ee432c6456ea288c changed coredata.xts()
to return an empty vector. The behavior of coredata.zoo() was the same
in zoo 1.5-6 (in 2009-05-21) as it is now.

Update coredata.xts() to only remove xts attributes, leaving other
attributes (e.g. dim, dimnames) untouched.

Fixes #223.