Do not coerce non-zero-width to zero-width's type
We were converting all xts objects to a common type when merging, and
we included the types of zero-width xts objects in our check. This is
a problem because zero-width objects created via the xts constructor
are numeric by default. So calling merge.xts() on any non-numeric xts
objects and a zero-width xts object returned a numeric result.

This manifested in the na.locf() test where xout is specified, because
na.approx.zoo() is called at one point. And that method does a merge
of the original object with an zero-width zoo object created via
zoo(,idx).

See #215.