Address potential issues found by rchk
The rchk tool found several potentially dangerous lines of code.

The xts_*ATTRIB() macro calls always allocated, since they coerce
from a pairlist to a list. But the result wasn't protected and was
available for GC.

The coerceVector() in the asReal(coerceVector()) calls was redundant.
asReal() coerces if needed.

Use PROTECT_WITH_INDEX and REPROTECT inside loops to avoid potential
protection stack overflows. Also use it in non_duplicates() to protect
the output of lengthgets() before returning the result.