Use by-column C code in na.locf
Loop over columns in C instead of in R. This should be faster because
the loop is faster in C, but it also avoids memory allocations that
happen at the R-level during subset-assignment.

This change required adding character support to the na_locf_col code
(now in the na_locf function), and to the firstNonNACol() function.