Avoid argument name partial match warnings
The parse.side() function inside .parseISO8601() returns a named
vector and the element containing the month value was named 'mon'.
This caused argument name partial match warnings because the
parse.side() output is passed to firstof() or lastof(), and the
argument for the month component of those functions is 'month'.

Change the parse.side() output to use 'month' instead of 'mon'. Thanks
to GitHub user gp2x for the report and the patch!

Fixes #231.