Reset gap counter for each column
Calling na.locf() on a multi-column xts object that is all NA except
for the first row did not carry the first row value forward. This was
caused by maxgap = nrow(x) and the gap counter not being reset after
looping over the first column. This would have also been a problem if
all but the last row were NA and na.locf() was called with fromLast =
FALSE.

Fixes #232.