Ensure make.index.unique result is always sorted
The functional change in this commit is adding eps when this is

    newindex_real[i] <= newindex_real[i-1]

instead of when this is true:

    index_real[i-1] == index_real[i]

This ensures the new index is always in increasing order, even
observation following a contiguous block of duplicate timestamps
less than the sum of the epsilons.

The memcpy ensures 'newindex' is equal to the current 'index'
the loop, which also means we longer need 'index_real'.

Fixes #241.