Add experimental C-based period apply
I became obsessed with improving the performance of period.apply()
after seeing its abysmal performance in this StackOverflow question:
https://stackoverflow.com/q/50062851/271616.

The C code works for the most part, but tests are needed (obviously).
The loop over the 'INDEX' (endpoints) was written very carefully to
avoid unnecessary allocations. For example, the call remains the same
for each period. The value bound to the crazily-named object changes
each iteration, which avoids creating a new call for each period.