Add structure and functions for both index types
The xts_index union provides a means to access any xts index, whether
it's stored as a double or an int. The xts_indices struct contains the
index for the 3 xts indices involved in the merge: result, x, and y.

These two data structures are then used by the set_index_from* and
compare_indexes* functions. The set_index_func and compare_func
function types provide an interface by which to call the
set_index_from* and compare_indexes* functions, respectively.

Determine and set the xts indices and the required functions before
starting the loop over the index. Since the functions and data handle
both double and integer indices, we can remove the branch for REALSXP
and INTSXP indices.