Automatically determine tick and grid locations
Both 'major.ticks' and 'grid.ticks.on' were set to "month" by default.
This didn't produce desirable output for most plots (anything other
than a year or two of daily data, really).

This fixes the immediate issue, but surfaces another bug. Now the grid
lines may not align with the marks and labels because of the infelicity
between axTicksByTime() and axTicksByTime2().

Fixes #246.