Restore support for both index types in window()
Moving the duplicate index value handling to C created an error when
the index type was integer. This was not caught in the unit tests.
Add a loop around the window/subset unit tests that run them on both
index types (double and integer). Add support for both index types to
the fill_window_dups_rev() C function.

Change the scalar '1' to '1L' in the pmax() call after findInterval()
to ensure that 'base_index' is not coerced to double from integer.

See #251.