Only use first tzone element in xts constructors
The 'tzone' attribute will usually be a single character string, but
it can be a character vector. For example, the POSIXlt 'tzone'
attribute has 3 elements: the actual timezone, and the abbreviations
for standard and daylight saving time.

Ensure we only store the first element in the 'tzone' attribute, in
order to avoid errors in downstream functions that always expect a
one-element 'tzone' attribute.

Fixes #105.