Fix if class(order.by) > 1
When class(order.by) has a length other than one, it will reach this
line but only the first element will be used.  We should make this
explicit and indeed must in case the environment variable _R_CHECK_LENGTH_1_LOGIC2_ is set, whereupon this line will emit an error
or warning.