Fix subset when index and i contain duplicates
The result of subsetting an object with 'i' that contains duplicates,
may be longer than 'i' and/or the object. But the code assumed the
subset result would never be longer than the object index. This worked
correctly prior to 1d707c57042203b7e53d7a8c5bd0ad173d9840d6, when the
fill_window_dups_rev() function was added.

Check for when the subset output length reaches the length of the
initially allocated result object. Increase the length of the result
object by twice the remaining elements in 'i' if we run out of space.

Fixes #275.