Fix 'if length > 1 logical' in .parseISO8601()
This will cause a failure in 'R CMD check' in R-3.6.0. This was
introduced in 16bed0feda52e71dc9648a9e4886bbd7da27a57f as part of #96
(prevent subset from returning all data when 'i' is not a valid date-
time).

This should only be an issue when 'intervals' is length-1 and the first
3 logical comparisons are TRUE. So we only need to check whether the
first element of 'intervals' is an empty string.

Also add unit tests for basic functionality, and add some fuzz tests
too.

Fixes #280.