Improve time-of-day subset performance
Use a slightly modified version of the cut_time_of_day() function
provided by StackOverflow user 3226167 in their answer here:
https://stackoverflow.com/a/39693699/271616

This improves performance by two orders of magnitude. The main
modification is to return an integer vector of locations instead of
a logical vector. This made usage inside of `[.xts` clearer.

Move the date-range subset to an "else" block, because we no longer
need that logic for time-of-day subsets.

Also remove .makeISO8601TT() since it is no longer used or needed.

Fixes #193.