Fix documentation for 'ticks' arguments
Details says that major.ticks, minor.ticks, and grid.ticks.on can have
periodicity-like values that end in "ly". These arguments are passed
to axTicksByTime(), which only recognizes values that end in "s".

Remove the comment about the values being based on the output from
periodicity(), because those values end in "ly".

Thanks to @harvey131 for the report and patch!

Fixes #291.