Only call as.character(dots) if necessary
We called as.character(dots) for all objects passed to merge(), even
if we had symbol names. This was incredibly expensive for large
objects.

Now we only call as.character() on the subset of dots that do not have
symbol names. This means we will often not call as.character() at all.

See #248.