Register as.zoo.xts() for zoo < 1.8-5
Kurt Hornik noticed xts:::as.zoo.xts() overrides zoo:::as.zoo.xts(),
which overwrites zoo's S3 registry entry when xts is loaded. The xts
method only exists because it didn't exist in zoo at the time. The
package that owns the class should define the conversion method, so
the as.zoo.*() methods should be in zoo.

Achim reported the error, and he noted the differences between the
zoo and xts versions of the method are very small. We agreed that the
xts version was slightly better, and the xts version should be moved
to zoo.

We can't remove xts:::as.zoo.xts() without causing test failures, so
we need to register xts:::as.zoo.xts() when zoo < 1.8-5 is installed
in order to avoid changing existing behavior.

Fixes #287.