Correct colnames and remove rownames for Ops.xts
The result of Ops.xts() on an xts and a matrix will have rownames if
the matrix has rownames (e.g. if the matrix was created via as.xts()).

Always remove rownames from the Ops.xts() result, and remove dimnames
entirely if the result does not have colnames.

Add tests that exercise subtraction. Need to add tests that exercise
the relational operators.

See #295.