Ensure Ops.xts results has index and no names
Calling Ops.xts() on a matrix and a xts object without a dim attribute
can result in an xts object without an index, and/or with a names
attribute.

Ensure there is never a names attribute, and that we always check if
the index exists.

Tests only exercise the subtraction operator. There should be tests
that use relational operators.

See #295.