Add all.equal.xts() method
We need an all.equal.xts() method to handle the case where all.equal()
is called with one xts object that has index attributes on the object
and one that only has them on the index. This can cause unit test
failures, and did in the "argo" package.

The function only removes the index attributes from either (or both)
objects before calling NextMethod(). I avoided the urge to check the
attributes attached to the object and index for consistency, and chose
what seemed to be a simpler approach.

See #245.