Handle index format attribute in .xts()
Check for .indexFORMAT and tformat passed to the constructor via
'...'. Warn that .indexFORMAT is deprecated, but still use it. We
will raise this to an error in a later version.

This means the special handling for Ops.xts() is now handled as part
of the more general case. It's also not strictly needed, because
Ops.xts() no longer passes .indexFORMAT to the constructor.

Remove these two elements from '...' before attaching the remaining
elements to the result (as user attributes).

See #245.