Handle index class attribute in .xts()
Check for .indexCLASS and tclass passed to the constructor via '...'.
Warn that .indexCLASS is deprecated, but still use it. We will raise
this to an error in a later version.

Also warn if the tclass attribute on the index is not the same as the
tclass argument default value ("POSIXct", "POSIXt"). Current behavior
over-rides the index tclass attribute with the default argument value.
This seems like a bug, but it may break existing code, so warn before
changing or converting to an error.

Remove these two elements from '...' before attaching the remaining
elements to the result (as user attributes).

Remove the indexClass() call from the checkXtsClass() test function,
since indexClass() is deprecated.

See #245.