Handle index timezone attribute in .xts()
Check for .indexTZ and tzone passed to the constructor via '...'. Warn
that .indexTZ is deprecated and will be ignored. This is consistent
with current behavior, but differs from the other two attributes that
can be over-ridden via arguments passed via '...'.

Remove these two elements from '...' before attaching the remaining
elements to the result (as user attributes).

Remove the indexTZ() call from the checkXtsTz() test function, since
indexTZ() is deprecated.

See #245.