Drop index attr, retain user attr passed via dots
We need to drop any index attributes (i.e deprecated .index* and
tformat) passed via '...', but retain any user attributes and standard
attributes like dim, dimnames, etc.

Note that tclass and tzone are not included in the dropped attributes
because they are formal arguments to the constructor, and therefore do
not appear in '...'.

See #245.