Unclass POSIXct for performance
The checks for NA in is.unsorted() and findInterval() are slower on
POSIXct vectors than on plain numeric vectors because ALTREP supports
faster NA checking on plain numeric vectors.

Leave the call to .toPOSIXct() before calling unclass() in order to
create any NA that may be needed due to timezones, or other date
weirdness. Continue to work on the plain numeric vector until the end
of the 'else' block.

Thanks to Harvey Smith for the report and suggestion.

Fixes #296.