Avoid multiple calls to dim()
Profiling revealed that calls to dim() were taking a non-trivial
amount of time in subset calls. This is unusual because dim() is a
primitive and very fast.

I noticed there are 3 calls to dim() near the top of the xts subset
function. One is explicit, to check for NULL. The other 2 are implicit
in the calls to nrow() and ncol().

Call dim() once and store the result. Then subset the output of dim()
to get the number of rows and columns.