Fix first/last on xtsible vector and period 'n'
The fix for #226 introduced a regression for the case when 'x' is an
xtsible vector and 'n' is a character period (e.g. "1 day"). This
occurred because the default methods for first() and last() call the
xts methods--on the *original* object--if 'x' can be coerced to xts.

For example, a call to
    first(.Date(1:3), "1 day")
would call
    first.xts(.Date(1:3), "1 day")

This was a problem because first.xts() and last.xts() assumed 'x'
would always have dims.

Fixes #303.