Ensure diff on logical xts preserves colnames
Column names were being dropped when diff() was called on an xts
object that had them.

Set dimnames in the .xts() constructor to match the incoming object.
Explicitly setting `dimnames(x) <- list(NULL, colnames(x))` caused
other tests for diff() on logical objects to fail, because it would
set dimnames to list(NULL, NULL) if the input did not have column
names.

Thanks to @sdittmar for the report!

Fixes #315.