Return zero-row matrix if input is zero-row
The default methods for first() and last() should return a zero-row
matrix if the input is a zero-length matrix. This makes them (almost)
consistent with head() and tail() in the utils package. (The only
difference is that utils::tail.matrix() adds rownames by default, and
last.default() will not.)

Also update the tests for vector inputs to check the output matches
the input, not the output of head()/tail(). This will avoid potential
test failures in the unlikely event that head()/tail() change in the
future.

See #309.